requirements.txt
gTTS==2.5.4
Flask==2.3.3
requests==2.31.0
google-cloud-texttospeech==2.16.3
